+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to the `dply` crate are documented in this file.
 ## Unreleased
 ### Changed 🔧
 * Update to Polars 0.29
+* Add `list_contains` to `filter` on list column values.
 
 ## 0.1.2 - 2023-05-09
 ### Changed 🔧